+++ libs/canvas-api-2/src/main/java/com/instructure/canvasapi2/managers/CourseManager.kt
@@ -371,12 +371,4 @@ object CourseManager {
         CourseAPI.getFirstPageCoursesWithGrades(adapter, callback, params)
     }
 
-    fun getFullCourseContent(courseId: Long, forceNetwork: Boolean, callback: StatusCallback<Course>) {
-        val adapter = RestBuilder(callback)
-        val params = RestParams(isForceReadFromNetwork = forceNetwork)
-        CourseAPI.getFullCourseContent(courseId, adapter, callback, params)
-    }
-
-    fun getFullCourseContentAsync(courseId: Long, forceNetwork: Boolean) = apiAsync<Course> { getFullCourseContent(courseId, forceNetwork, it) }
-
 }
