+++ libs/pandautils/src/main/AndroidManifest.xml
@@ -19,6 +19,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.instructure.pandautils" >
 
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <application>
         <activity
             android:name="com.instructure.pandautils.utils.AvatarCropActivity"
